requirements.txt
minecraft_launcher_lib===6.5
simple-term-menu===1.6.6
PyYAML===6.0.2
prettytable===3.12.0